backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
openai==1.45.0
pydantic==2.9.1
